Refactor table assertion to reusable helper
